move encrypt command under /actions
